% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{apply_transform}
\alias{apply_transform}
\title{Apply transformation matrix to a variable}
\usage{
apply_transform(y, H, is_lbmgls)
}
\arguments{
\item{y}{n x 1 vector}

\item{H}{n x n transformation matrix}

\item{is_lbmgls}{Whether to mean-center the result}
}
\value{
n x 1 transformed vector
}
\description{
hy = H * y. For lbmgls, additionally mean-center the result.
}
