% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_missing}
\alias{check_missing}
\title{Check for missing values in a matrix and report}
\usage{
check_missing(mat, name = "data")
}
\arguments{
\item{mat}{A numeric matrix.}

\item{name}{Character; name for error messages.}
}
\description{
Check for missing values in a matrix and report
}
\keyword{internal}
