% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cluster_matrix}
\alias{cluster_matrix}
\title{Compute cluster transformation matrix}
\usage{
cluster_matrix(cluster)
}
\arguments{
\item{cluster}{n x 1 vector of integer cluster IDs}
}
\value{
n x n cluster transformation matrix
}
\description{
clust_mat = I - (clust_id == clust_id') / rowsum(clust_id == clust_id')
}
