% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{double_center}
\alias{double_center}
\title{Double-center a matrix (subtract row means, then column means)}
\usage{
double_center(mat)
}
\arguments{
\item{mat}{n x n matrix}
}
\value{
Double-centered matrix
}
\description{
Matches the Stata demeanmat() function.
}
