% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{euclidean_distances}
\alias{euclidean_distances}
\title{Compute Euclidean distance matrix}
\usage{
euclidean_distances(coords)
}
\arguments{
\item{coords}{n x d matrix of coordinates}
}
\value{
n x n Euclidean distance matrix
}
\description{
Compute Euclidean distance matrix
}
