% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_core.R
\name{find_ha_parm_i0}
\alias{find_ha_parm_i0}
\title{Find ha_parm g for I(0) test}
\usage{
find_ha_parm_i0(om_ho, om_i0, om_bm, emat, maxiter = 20)
}
\arguments{
\item{om_ho}{q x q null covariance}

\item{om_i0}{q x q white noise covariance}

\item{om_bm}{q x q BM covariance}

\item{emat}{q x nrep standard normal draws}
}
\value{
ha_parm scalar
}
\description{
Find ha_parm g for I(0) test
}
\keyword{internal}
