% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_core.R
\name{find_ha_parm_i1}
\alias{find_ha_parm_i1}
\title{Find ha_parm for I(1) test (R implementation)}
\usage{
find_ha_parm_i1(om_ho, distmat, Rmat, emat, maxiter = 20)
}
\arguments{
\item{om_ho}{q x q null covariance}

\item{distmat}{n x n distance matrix}

\item{Rmat}{n x q eigenvector matrix}

\item{emat}{q x nrep standard normal draws}
}
\value{
ha_parm scalar
}
\description{
Find ha_parm for I(1) test (R implementation)
}
\keyword{internal}
