% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_R}
\alias{get_R}
\title{Get top q eigenvectors of a symmetric matrix}
\usage{
get_R(sigma, q)
}
\arguments{
\item{sigma}{n x n symmetric matrix}

\item{q}{Number of eigenvectors to extract}
}
\value{
n x q matrix of top eigenvectors
}
\description{
Computes eigendecomposition, sorts eigenvalues descending,
returns first q eigenvectors.
Matches the Stata get_R() function.
}
