% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getcbar}
\alias{getcbar}
\title{Find c such that mean(exp(-c * lower_tri(distmat))) equals rhobar}
\usage{
getcbar(rhobar, distmat)
}
\arguments{
\item{rhobar}{Target correlation}

\item{distmat}{n x n distance matrix}
}
\value{
The constant cbar
}
\description{
Uses exponential bracketing followed by geometric bisection.
Matches the Stata getcbar() function exactly.
}
