% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_core.R
\name{getcbar_r}
\alias{getcbar_r}
\title{Get cbar: find c such that mean(exp(-c * lower_tri(dist))) = rhobar}
\usage{
getcbar_r(rhobar, distmat)
}
\arguments{
\item{rhobar}{Target correlation}

\item{distmat}{n x n distance matrix}
}
\value{
cbar
}
\description{
Calls the C++ implementation (which works fine).
}
\keyword{internal}
