% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getpow_qf}
\alias{getpow_qf}
\title{Compute power of quadratic form test}
\usage{
getpow_qf(om0_in, om1_in, e)
}
\arguments{
\item{e}{q x nrep matrix of standard normal draws}

\item{om0}{Null hypothesis covariance matrix (q x q)}

\item{om1}{Alternative hypothesis covariance matrix (q x q)}
}
\value{
Power (scalar between 0 and 1)
}
\description{
Compares quadratic forms under null (om0) vs alternative (om1) covariance.
Matches the Stata getpow_qf() function.
}
