% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_core.R
\name{getpow_qf_r}
\alias{getpow_qf_r}
\title{Compute power of quadratic form test (R implementation)}
\usage{
getpow_qf_r(om0, om1, emat)
}
\arguments{
\item{om0}{Null hypothesis q x q covariance matrix}

\item{om1}{Alternative q x q covariance matrix}

\item{emat}{q x nrep standard normal draws}
}
\value{
Power scalar
}
\description{
Compute power of quadratic form test (R implementation)
}
\keyword{internal}
