% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{haversine_distances}
\alias{haversine_distances}
\title{Compute Haversine great-circle distance matrix}
\usage{
haversine_distances(coords)
}
\arguments{
\item{coords}{n x 2 matrix with latitude in column 0, longitude in column 1}
}
\value{
n x n Haversine distance matrix (scaled by 1/pi)
}
\description{
Distances are divided by pi to match Stata convention (fraction of
semi-circle, so max distance = 1).
}
