% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{iso_matrix}
\alias{iso_matrix}
\title{Compute isotropic transformation matrix}
\usage{
iso_matrix(s, radius, latlong)
}
\arguments{
\item{s}{n x d matrix of locations}

\item{radius}{Radius for differencing}

\item{latlong}{Whether coordinates are latitude/longitude}
}
\value{
n x n isotropic transformation matrix
}
\description{
iso_mat = I - dist_below_b / rowsum(dist_below_b)
where dist_below_b indicates distances <= radius
}
