% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lbm_gls_matrix}
\alias{lbm_gls_matrix}
\title{Compute LBM-GLS transformation matrix}
\usage{
lbm_gls_matrix(distmat)
}
\arguments{
\item{distmat}{n x n distance matrix}
}
\value{
n x n LBM-GLS transformation matrix
}
\description{
From sigma_lbm_dm eigendecomposition, keep eigenvalues > 1e-10,
return V * diag(1/sqrt(eval)) * V'.
}
