% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lvech}
\alias{lvech}
\title{Half-vectorization: extract lower triangular elements (excluding diagonal)}
\usage{
lvech(S)
}
\arguments{
\item{S}{n x n matrix}
}
\value{
Column vector of lower-triangular elements
}
\description{
Matches the Stata lvech() function.
}
