% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nn_matrix}
\alias{nn_matrix}
\title{Compute nearest-neighbor transformation matrix}
\usage{
nn_matrix(s, latlong)
}
\arguments{
\item{s}{n x d matrix of locations}

\item{latlong}{Whether coordinates are latitude/longitude}
}
\value{
n x n NN transformation matrix
}
\description{
NN_mat = I - rowmins_mat, where each row has 1/k for the k nearest neighbors.
}
