% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{normalize_distances}
\alias{normalize_distances}
\title{Normalize distance matrix so maximum distance is 1}
\usage{
normalize_distances(distmat)
}
\arguments{
\item{distmat}{n x n distance matrix}
}
\value{
Normalized distance matrix
}
\description{
Normalize distance matrix so maximum distance is 1
}
