% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sigma_dm}
\alias{sigma_dm}
\title{Compute demeaned spatial covariance with exponential decay}
\usage{
sigma_dm(distmat, c)
}
\arguments{
\item{distmat}{n x n distance matrix}

\item{c}{Decay parameter}
}
\value{
Double-centered exponential covariance matrix
}
\description{
sigma = exp(-c * distmat), then double-center.
}
