% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sigma_lbm}
\alias{sigma_lbm}
\title{Compute LBM covariance matrix from distance matrix}
\usage{
sigma_lbm(distmat)
}
\arguments{
\item{distmat}{n x n distance matrix}
}
\value{
n x n LBM covariance matrix
}
\description{
sigma_lbm = 0.5 * (J * dist\link{,1}' + dist\link{1,} * J' - dist)
Uses the first location as the origin.
}
