% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sigma_lbm_dm}
\alias{sigma_lbm_dm}
\title{Compute demeaned LBM covariance matrix}
\usage{
sigma_lbm_dm(distmat)
}
\arguments{
\item{distmat}{n x n distance matrix}
}
\value{
Double-centered LBM covariance matrix
}
\description{
sigma_lbm_dm = demeanmat(sigma_lbm(distmat))
}
