% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sigma_residual}
\alias{sigma_residual}
\title{Compute residual-projected spatial covariance}
\usage{
sigma_residual(distmat, c, M)
}
\arguments{
\item{distmat}{n x n distance matrix}

\item{c}{Decay parameter}

\item{M}{n x n projection matrix (I - X(X'X)^{-1}X')}
}
\value{
Residual-projected covariance matrix
}
\description{
sigma = exp(-c * distmat), then sigma_dm = M * sigma * M'.
}
