% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_core.R
\name{spatial_i0_core}
\alias{spatial_i0_core}
\title{Spatial I(0) test (R implementation)}
\usage{
spatial_i0_core(Y, distmat, emat, q)
}
\arguments{
\item{Y}{n x 1 vector}

\item{distmat}{n x n normalized distance matrix}

\item{emat}{q x nrep standard normal draws}

\item{q}{Number of low-frequency components}
}
\value{
List with LR, pvalue, cvalue, ha_parm, rho_grid, pvalue_mat, cvalue_mat
}
\description{
Spatial I(0) test (R implementation)
}
\keyword{internal}
