% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spatial_i0_test}
\alias{spatial_i0_test}
\title{Spatial I(0) unit root test}
\usage{
spatial_i0_test(Y, distmat, emat, q)
}
\arguments{
\item{Y}{n x 1 vector of observations}

\item{distmat}{n x n normalized distance matrix}

\item{emat}{q x nrep matrix of standard normal draws}

\item{q}{Number of low-frequency components}
}
\value{
List with LR, pvalue, cvalue, ha_parm, rho_grid, pvalue_mat, cvalue_mat
}
\description{
Tests H0: no spatial unit root (I(0)) vs H1: spatial unit root.
Uses a grid of rho values for the null hypothesis.
}
