% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spatial_i0_test_residual}
\alias{spatial_i0_test_residual}
\title{Spatial I(0) residual unit root test}
\usage{
spatial_i0_test_residual(Y, X_in, distmat, emat, q)
}
\arguments{
\item{Y}{n x 1 response vector}

\item{X_in}{n x k design matrix (including intercept)}

\item{distmat}{n x n normalized distance matrix}

\item{emat}{q x nrep matrix of standard normal draws}

\item{q}{Number of low-frequency components}
}
\value{
List with LR, pvalue, cvalue, ha_parm, rho_grid, pvalue_mat, cvalue_mat
}
\description{
Tests residuals from regression Y on X for no spatial unit root (I(0)).
}
