% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spatial_i1_test}
\alias{spatial_i1_test}
\title{Spatial I(1) unit root test}
\usage{
spatial_i1_test(Y, distmat, emat, q)
}
\arguments{
\item{Y}{n x 1 vector of observations}

\item{distmat}{n x n normalized distance matrix}

\item{emat}{q x nrep matrix of standard normal draws}

\item{q}{Number of low-frequency components}
}
\value{
List with LR, pvalue, ha_parm, cv_vec
}
\description{
Tests H0: spatial unit root (I(1)) vs H1: not I(1).
}
