% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spatial_i1_test_residual}
\alias{spatial_i1_test_residual}
\title{Spatial I(1) residual unit root test}
\usage{
spatial_i1_test_residual(Y, X_in, distmat, emat, q)
}
\arguments{
\item{Y}{n x 1 response vector}

\item{X_in}{n x k design matrix (including intercept)}

\item{distmat}{n x n normalized distance matrix}

\item{emat}{q x nrep matrix of standard normal draws}

\item{q}{Number of low-frequency components}
}
\value{
List with LR, pvalue, ha_parm, cv_vec
}
\description{
Tests residuals from regression Y on X for spatial unit root.
}
