% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_core.R
\name{spatial_persistence_core}
\alias{spatial_persistence_core}
\title{Spatial half-life confidence set (R implementation)}
\usage{
spatial_persistence_core(Z, distmat, emat, level)
}
\description{
Spatial half-life confidence set (R implementation)
}
